fix: bring topology updates back
crontab and with that the topology functions used to be set up from within the info pane. Thinking about it now makes me wonder why I moved it there in the first place - it was clearly a mistake.

Since the info pane is not used anymore the topology functions are also not triggered anymore.

This commit brings it back as its essential for the relays as long as we don't have the p2p module activated.