build bech32 on debian bookwork
building on the previous haskell-debian image was failing so upgrading was
needed. In this newer image stack doesn't seem to be available by
default, so removing the initial upgrade command was needed. Not
providing a version when setting up stack was leading to compilation
errors as well, so I had to fix the version to 9.8.4